+++ exportplan/core/data.py
@@ -73,6 +73,7 @@ LESSONS = {
     ],
 }
 
+
 SECTIONS = OrderedDict(
     {
         slugify(ABOUT_YOUR_BUSINESS): {
