+++ exportplan/core/helpers.py
@@ -4,8 +4,8 @@ from iso3166 import countries_by_alpha3
 from core import models
 from core.templatetags.content_tags import format_timedelta
 from directory_api_client import api_client
-from directory_constants import choices
-from exportplan import data, serializers
+from exportplan.core.processor import ExportPlanProcessor
+from . import data
 
 
 def create_export_plan(sso_session_id, exportplan_data):
@@ -273,99 +273,16 @@ def values_to_labels(values, choices):
 
 
 def get_export_plan_pdf_context(request):
+    processor = ExportPlanProcessor(request.user.export_plan.data)
     context = {
         'host_url': ('https://' if request.is_secure() else 'http://') + request.get_host(),
         'export_plan': request.user.export_plan.data,
         'user': request.user,
         'sections': data.SECTION_TITLES,
-        'calculated_pricing': request.user.export_plan.calculated_cost_pricing(),
-        'total_funding': request.user.export_plan.calculate_total_funding(),
+        'calculated_pricing': processor.calculated_cost_pricing(),
+        'total_funding': processor.calculate_total_funding(),
         'getting_paid_payment_method_label': request.user.export_plan.getting_paid_payment_method_label,
         'getting_paid_incoterms_transport_label': request.user.export_plan.getting_paid_incoterms_transport_label,
     }
 
     return context
-
-
-class ExportPlanParser:
-    """
-    Parse the export plan details provided by directory-api's exportplan
-    serializer
-
-    """
-
-    PAYMENT_METHOD_OPTIONS = dict(choices.PAYMENT_METHOD_OPTIONS)
-    ALL_TRANSPORT_OPTIONS = dict(choices.TRANSPORT_OPTIONS + choices.WATER_TRANSPORT_OPTIONS)
-
-    def __init__(self, data):
-        self.data = data
-
-    def __bool__(self):
-        return bool(self.data)
-
-    @property
-    def export_country_name(self):
-        if self.data.get('export_countries'):
-            return self.data['export_countries'][0]['country_name']
-
-    @property
-    def export_country_code(self):
-        if self.data.get('export_countries'):
-            return self.data['export_countries'][0]['country_iso2_code']
-
-    @property
-    def export_commodity_code(self):
-        if self.data.get('export_commodity_codes'):
-            return self.data['export_commodity_codes'][0]['commodity_code']
-
-    def build_current_url(self, slug):
-        current_url = data.SECTIONS[slug]
-        current_url.pop('country_required', None)
-        current_url.pop('product_required', None)
-        if slug in data.COUNTRY_REQUIRED:
-            if not self.data.get('export_countries') or len(self.data['export_countries']) == 0:
-                current_url['country_required'] = True
-        if slug in data.PRODUCT_REQUIRED:
-            if not self.data.get('export_commodity_codes') or len(self.data['export_commodity_codes']) == 0:
-                current_url['product_required'] = True
-        current_url['is_complete'] = self.data.get('ui_progress', {}).get(slug, {}).get('is_complete', False)
-        return current_url
-
-    def build_export_plan_sections(self):
-        sections = data.SECTIONS
-        for slug, values in sections.items():
-            values['is_complete'] = self.data.get('ui_progress', {}).get(slug, {}).get('is_complete', False)
-        return list(sections.values())
-
-    def calculated_cost_pricing(self):
-        calculated_pricing = serializers.ExportPlanSerializer(data=self.data).calculate_cost_pricing
-        return {'calculated_cost_pricing': calculated_pricing}
-
-    def calculate_total_funding(self):
-        total_funding = serializers.ExportPlanSerializer(data=self.data).calculate_total_funding
-        return {'calculated_total_funding': total_funding}
-
-    def calculate_ep_progress(self):
-        progress_items = self.data.get('ui_progress', {})
-        completed = [True for v in progress_items.values() if v.get('is_complete')]
-        return {
-            'export_plan_progress': {
-                'sections_completed': len(completed),
-                'sections_total': len(data.SECTION_SLUGS),
-                'percentage_completed': len(completed) / len(data.SECTION_SLUGS) if len(completed) > 0 else 0,
-            }
-        }
-
-    @property
-    def getting_paid_payment_method_label(self):
-        return values_to_labels(
-            values=self.data.get('getting_paid', {}).get('payment_method', {}).get('methods') or [],
-            choices=self.PAYMENT_METHOD_OPTIONS,
-        )
-
-    @property
-    def getting_paid_incoterms_transport_label(self):
-        return values_to_labels(
-            values=[self.data.get('getting_paid', {}).get('incoterms', {}).get('transport')] or [],
-            choices=self.ALL_TRANSPORT_OPTIONS,
-        )
