+++ exportplan/models.py
@@ -3,6 +3,7 @@ from great_components.mixins import GA360Mixin
 from wagtail.core.models import Page
 
 from core import cms_slugs, mixins
+from exportplan.core.processor import ExportPlanProcessor
 
 
 class ExportPlanDashboardPage(
@@ -18,10 +19,11 @@ class ExportPlanDashboardPage(
     template = 'exportplan/dashboard_page.html'
 
     def get_context(self, request):
+        processor = ExportPlanProcessor(request.user.export_plan.data)
         request.user.set_page_view(cms_slugs.EXPORT_PLAN_DASHBOARD_URL)
         context = super().get_context(request)
-        context['sections'] = request.user.export_plan.build_export_plan_sections()
-        context['export_plan_progress'] = request.user.export_plan.calculate_ep_progress()
+        context['sections'] = processor.build_export_plan_sections()
+        context['export_plan_progress'] = processor.calculate_ep_progress()
 
         self.set_ga360_payload(
             page_id=self.id,
