+++ tests/unit/exportplan/test_ajax.py
@@ -5,7 +5,7 @@ import pytest
 from django.urls import reverse
 from freezegun import freeze_time
 
-from exportplan import helpers
+from exportplan.core import helpers
 
 
 @pytest.mark.django_db
