+++ tests/unit/exportplan/test_helpers.py
@@ -3,7 +3,8 @@ from unittest import mock
 import pytest
 
 from directory_api_client import api_client
-from exportplan import data, helpers
+from exportplan.core import helpers
+from exportplan.core.parsers import ExportPlanParser
 from tests.helpers import create_response
 
 
@@ -144,7 +145,7 @@ def test_serialize_exportplan_data_with_country_expertise(user, mock_get_company
 
 
 def test_export_plan_parser_getting_paid_labels(user, export_plan_data):
-    export_plan_parser = helpers.ExportPlanParser(export_plan_data)
+    export_plan_parser = ExportPlanParser(export_plan_data)
     assert export_plan_parser.getting_paid_incoterms_transport_label == 'Ex Works (EXW)'
     assert export_plan_parser.getting_paid_payment_method_label == 'Credit or debit card payments, Merchant services'
 
@@ -509,104 +510,9 @@ def test_update_ui_options_target_ages_not_required(mock_update_export_plan, exp
     assert mock_update_export_plan.call_count == 0
 
 
-@pytest.mark.parametrize(
-    'ui_progress_data, complete, percentage_complete',
-    [
-        [{}, 0, 0],
-        [{'a': {}}, 0, 0],
-        [{'a': {'is_complete': False}}, 0, 0],
-        [{'a': {'is_complete': True}}, 1, 0.1],
-        [{'b': {'is_complete': True}, 'c': {'is_complete': True}}, 2, 0.2],
-    ],
-)
-@mock.patch.object(helpers, 'get_exportplan')
-def test_export_plan_parser_calculate_ep_progress(mock_get_exportplan, ui_progress_data, complete, percentage_complete):
-    export_plan_data = {'ui_progress': ui_progress_data}
-    mock_get_exportplan.return_value = export_plan_data
-    ep_progress = helpers.ExportPlanParser(export_plan_data).calculate_ep_progress()['export_plan_progress']
-    assert ep_progress['sections_total'] == len(data.SECTION_SLUGS)
-    assert ep_progress['sections_completed'] == complete
-    assert ep_progress['percentage_completed'] == percentage_complete
-
-
-def test_export_plan_parser_build_export_plan_sections(export_plan_data):
-    sections = helpers.ExportPlanParser(export_plan_data).build_export_plan_sections()
-    assert sections[0]['is_complete'] is True
-    assert sections[1]['is_complete'] is False
-
-
-def test_export_plan_parser_calculated_cost_pricing(cost_pricing_data):
-    pricing_data = helpers.ExportPlanParser(cost_pricing_data).calculated_cost_pricing()
-    assert pricing_data == {
-        'calculated_cost_pricing': {
-            'total_direct_costs': '15.00',
-            'total_overhead_costs': '1355.00',
-            'profit_per_unit': '6.00',
-            'potential_total_profit': '132.00',
-            'gross_price_per_unit': '42.36',
-            'total_export_costs': '1685.00',
-            'estimated_costs_per_unit': '76.59',
-        }
-    }
-
-
-@pytest.mark.parametrize(
-    'export_plan_data, expected',
-    [
-        [{'export_countries': [{'country_name': 'Netherlands', 'country_iso2_code': 'NL'}]}, None],
-        [{'export_countries': []}, True],
-        [{'export_countries': None}, True],
-    ],
-)
-def test_export_plan_parser_get_current_url_country_required(export_plan_data, expected):
-    current_url = helpers.ExportPlanParser(export_plan_data).build_current_url('target-markets-research')
-    assert current_url.get('country_required') == expected
-
-
-def test_export_plan_parser_get_current_url_country_required_not_in_check():
-    export_plan_data = {'export_countries': []}
-    current_url = helpers.ExportPlanParser(export_plan_data).build_current_url('about-your-business')
-    assert current_url.get('country_required') is None
-
-
-@pytest.mark.parametrize(
-    'export_plan_data, expected',
-    [
-        [{'export_commodity_codes': [{'commodity_code': '220850', 'commodity_name': 'Gin'}]}, None],
-        [{'export_commodity_codes': []}, True],
-        [{'export_commodity_codes': None}, True],
-    ],
-)
-def test_export_plan_parser_get_current_url_product_required(export_plan_data, expected):
-    current_url = helpers.ExportPlanParser(export_plan_data).build_current_url('target-markets-research')
-    assert current_url.get('product_required') == expected
-
-
-def test_export_plan_parser_get_current_url_product_required_not_in_check():
-    export_plan_data = {'export_commodity_codes': []}
-    current_url = helpers.ExportPlanParser(export_plan_data).build_current_url('about-your-business')
-    assert current_url.get('product_required') is None
-
-
-@pytest.mark.parametrize(
-    'ui_progress_data, expected',
-    [
-        [{}, False],
-        [{'target-markets': {'is_complete': True}}, False],
-        [{'about-your-business': {}}, False],
-        [{'about-your-business': {'is_complete': False}}, False],
-        [{'about-your-business': {'is_complete': True}, 'Target-markets': {'is_complete': True}}, True],
-    ],
-)
-def test_export_plan_parser_get_current_url_progress(ui_progress_data, expected):
-    export_plan_data = {'ui_progress': ui_progress_data}
-    current_url = helpers.ExportPlanParser(export_plan_data).build_current_url('about-your-business')
-    assert current_url.get('is_complete') is expected
-
-
 def test_export_plan_parser(export_plan_data):
 
-    ep_parser = helpers.ExportPlanParser(export_plan_data)
+    ep_parser = ExportPlanParser(export_plan_data)
     assert ep_parser.data == export_plan_data
     assert ep_parser.export_country_name == export_plan_data['export_countries'][0]['country_name']
     assert ep_parser.export_commodity_code == export_plan_data['export_commodity_codes'][0]['commodity_code']
